+++ Sources/ZUtils.swift
@@ -41,97 +41,3 @@ public func cleanLogRecords(_ context: NSManagedObjectContext, keepSince: Date) 
     try context.deleter(entityName: "ZRoutine",
                         predicate: NSPredicate(format: "zRoutineRuns.@count == 0"))
 }
-
-internal enum ZType {
-    case zroutine
-    case zroutinerun
-    case zexercise
-    case zexerciserun
-}
-
-internal typealias ZTypeObjectIDs = [ZType: [NSManagedObjectID]]
-
-/// Deep copy of all routines and their children from the source store to specified destination store
-/// Returning list of the objectIDs of the records copied FROM the SOURCE store.
-/// Does not delete any records.
-/// Does NOT save context.
-internal func deepCopy(_ context: NSManagedObjectContext,
-                       fromStore srcStore: NSPersistentStore,
-                       toStore dstStore: NSPersistentStore) throws -> ZTypeObjectIDs
-{
-    var copiedObjects = ZTypeObjectIDs()
-
-    func append(_ ztype: ZType, _ objectID: NSManagedObjectID) {
-        copiedObjects[ztype, default: [NSManagedObjectID]()]
-            .append(objectID)
-    }
-
-    try context.fetcher(ZRoutine.self, inStore: srcStore) { sRoutine in
-
-        let dRoutine = try sRoutine.shallowCopy(context, toStore: dstStore)
-
-        let routinePred = NSPredicate(format: "zRoutine = %@", sRoutine)
-
-        try context.fetcher(ZRoutineRun.self, predicate: routinePred, inStore: srcStore) { sRoutineRun in
-
-            _ = try sRoutineRun.shallowCopy(context, dstRoutine: dRoutine, toStore: dstStore)
-
-            append(.zroutinerun, sRoutineRun.objectID)
-            print("Copied zRoutineRun \(sRoutine.wrappedName) startedAt=\(String(describing: sRoutineRun.startedAt))")
-            return true
-        }
-
-        try context.fetcher(ZExercise.self, predicate: routinePred, inStore: srcStore) { sExercise in
-
-            let dExercise = try sExercise.shallowCopy(context, dstRoutine: dRoutine, toStore: dstStore)
-
-            let exercisePred = NSPredicate(format: "zExercise = %@", sExercise)
-
-            try context.fetcher(ZExerciseRun.self, predicate: exercisePred, inStore: srcStore) { sExerciseRun in
-
-                _ = try sExerciseRun.shallowCopy(context, dstExercise: dExercise, toStore: dstStore)
-
-                append(.zexerciserun, sExerciseRun.objectID)
-                print("Copied zExerciseRun \(sExercise.wrappedName) completedAt=\(String(describing: sExerciseRun.completedAt))")
-                return true
-            }
-
-            append(.zexercise, sExercise.objectID)
-            print("Copied zExercise \(sExercise.wrappedName)")
-
-            return true
-        }
-
-        append(.zroutine, sRoutine.objectID)
-        print("Copied zRoutine \(sRoutine.wrappedName)")
-
-        return true
-    }
-
-    return copiedObjects
-}
-
-/// Transfers all 'Z' records in .main store to .archive store.
-/// NOTE Does save context
-public func transferToArchive(_ context: NSManagedObjectContext) throws {
-    print("\(#function)")
-    guard let mainURL = PersistenceManager.stores[.main]?.url,
-          let archiveURL = PersistenceManager.stores[.archive]?.url,
-          let psc = context.persistentStoreCoordinator,
-          let mainStore = psc.persistentStore(for: mainURL),
-          let archiveStore = psc.persistentStore(for: archiveURL)
-    else {
-        throw DataError.transferError(msg: "Unexpected store configuration.")
-    }
-
-    do {
-        let srcObjectIdDict = try deepCopy(context, fromStore: mainStore, toStore: archiveStore)
-        if srcObjectIdDict.count > 0 {
-            try srcObjectIdDict.values.forEach {
-                try context.deleter(objectIDs: $0)
-            }
-        }
-    } catch {
-        throw DataError.transferError(msg: error.localizedDescription)
-    }
-}
